+++ chain/actors/builtin/miner/miner.go
@@ -211,6 +211,7 @@ type Deadline interface {
 
 	PartitionsChanged(Deadline) (bool, error)
 	DisputableProofCount() (uint64, error)
+	DailyFee() abi.TokenAmount
 }
 
 type Partition interface {