+++ chain/actors/builtin/miner/v0.go
@@ -480,6 +480,10 @@ func (d *deadline0) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline0) DailyFee() abi.TokenAmount {
+	return big.Zero()
+}
+
 func (p *partition0) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }