+++ chain/actors/builtin/miner/v12.go
@@ -515,6 +515,10 @@ func (d *deadline12) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline12) DailyFee() abi.TokenAmount {
+	return big.Zero()
+}
+
 func (p *partition12) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }