+++ chain/actors/builtin/miner/v13.go
@@ -515,6 +515,10 @@ func (d *deadline13) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline13) DailyFee() abi.TokenAmount {
+	return big.Zero()
+}
+
 func (p *partition13) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }