+++ chain/actors/builtin/miner/v14.go
@@ -515,6 +515,10 @@ func (d *deadline14) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline14) DailyFee() abi.TokenAmount {
+	return big.Zero()
+}
+
 func (p *partition14) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }