+++ chain/actors/builtin/miner/v16.go
@@ -13,6 +13,7 @@ import (
 	rle "github.com/filecoin-project/go-bitfield/rle"
 	"github.com/filecoin-project/go-state-types/abi"
 	actorstypes "github.com/filecoin-project/go-state-types/actors"
+	"github.com/filecoin-project/go-state-types/big"
 	builtin16 "github.com/filecoin-project/go-state-types/builtin"
 	miner16 "github.com/filecoin-project/go-state-types/builtin/v16/miner"
 	adt16 "github.com/filecoin-project/go-state-types/builtin/v16/util/adt"
@@ -514,6 +515,13 @@ func (d *deadline16) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline16) DailyFee() abi.TokenAmount {
+	if d.Deadline.DailyFee.Int != nil {
+		return d.Deadline.DailyFee
+	}
+	return big.Zero()
+}
+
 func (p *partition16) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }
@@ -531,6 +539,10 @@ func (p *partition16) UnprovenSectors() (bitfield.BitField, error) {
 }
 
 func fromV16SectorOnChainInfo(v16 miner16.SectorOnChainInfo) SectorOnChainInfo {
+	dailyFee := v16.DailyFee
+	if dailyFee.Int == nil {
+		dailyFee = big.Zero()
+	}
 	info := SectorOnChainInfo{
 		SectorNumber:          v16.SectorNumber,
 		SealProof:             v16.SealProof,
@@ -547,7 +559,7 @@ func fromV16SectorOnChainInfo(v16 miner16.SectorOnChainInfo) SectorOnChainInfo {
 		PowerBaseEpoch:        v16.PowerBaseEpoch,
 		ReplacedDayReward:     v16.ReplacedDayReward,
 		Flags:                 SectorOnChainInfoFlags(v16.Flags),
-		DailyFee:              v16.DailyFee,
+		DailyFee:              dailyFee,
 	}
 	return info
 }