+++ chain/actors/builtin/miner/v3.go
@@ -512,6 +512,10 @@ func (d *deadline3) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline3) DailyFee() abi.TokenAmount {
+	return big.Zero()
+}
+
 func (p *partition3) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }