+++ chain/actors/builtin/miner/v5.go
@@ -512,6 +512,10 @@ func (d *deadline5) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline5) DailyFee() abi.TokenAmount {
+	return big.Zero()
+}
+
 func (p *partition5) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }