+++ chain/actors/builtin/miner/v6.go
@@ -512,6 +512,10 @@ func (d *deadline6) DisputableProofCount() (uint64, error) {
 
 }
 
+func (d *deadline6) DailyFee() abi.TokenAmount {
+	return big.Zero()
+}
+
 func (p *partition6) AllSectors() (bitfield.BitField, error) {
 	return p.Partition.Sectors, nil
 }