+++ itests/manual_onboarding_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ipfs/go-cid"
 	"github.com/stretchr/testify/require"
 
 	"github.com/filecoin-project/go-state-types/abi"
@@ -87,7 +88,7 @@ func TestManualSectorOnboarding(t *testing.T) {
 			minerC.AssertNoPower()
 
 			// ---- Miner B onboards a CC sector
-			bSectors := minerB.OnboardSectors(sealProofType, false, 1)
+			bSectors, _ := minerB.OnboardSectors(sealProofType, []kit.SectorManifest{{}})
 			req.Len(bSectors, 1)
 			// Miner B should still not have power as power can only be gained after sector is activated i.e. the first WindowPost is submitted for it
 			minerB.AssertNoPower()
@@ -95,7 +96,7 @@ func TestManualSectorOnboarding(t *testing.T) {
 			blockMiner.WatchMinerForPost(minerB.ActorAddr)
 
 			// --- Miner C onboards sector with data/pieces
-			cSectors := minerC.OnboardSectors(sealProofType, true, 1)
+			cSectors, _ := minerC.OnboardSectors(sealProofType, []kit.SectorManifest{kit.SectorWithRandPiece()})
 			// Miner C should still not have power as power can only be gained after sector is activated i.e. the first WindowPost is submitted for it
 			minerC.AssertNoPower()
 			// Ensure that the block miner checks for and waits for posts during the appropriate proving window from our new miner with a sector
@@ -106,7 +107,7 @@ func TestManualSectorOnboarding(t *testing.T) {
 			minerC.WaitTillActivatedAndAssertPower(cSectors, uint64(defaultSectorSize), uint64(defaultSectorSize))
 
 			// Miner B has activated the CC sector -> upgrade it with snapdeals
-			_ = minerB.SnapDeal(bSectors[0])
+			_, _ = minerB.SnapDeal(bSectors[0], kit.SectorManifest{Piece: cid.MustParse("baga6ea4seaqlhznlutptgfwhffupyer6txswamerq5fc2jlwf2lys2mm5jtiaeq")})
 		})
 	}
 }